added mass props

+++ tests/fibergen.py
@@ -13,6 +13,8 @@ class FiberGen:
         self.simbox_ysize = 4.0
         self.simbox_zsize = 4.0
         self.mc_field_value = 1.1
+        self.fc_mutate_weight = 1.0
+        self.fc_normal_weight = 0.0
         self.num_additional_steps = 5
         self.periodic = True
         self.show_viewer = True
@@ -20,6 +22,8 @@ class FiberGen:
         self.show_axes = False
         self.num_threads = 1
         self.capsules = []
+        self.keepjson = False
+        self.field_power = 2.0
 
     def append_capsule(self, ca, cb, cr):
         self.capsules.append(Capsule(ca, cb, cr))
@@ -37,11 +41,14 @@ class FiberGen:
         jdict["mc_field_value"] = self.mc_field_value
         jdict["num_additional_steps"] = self.num_additional_steps
         jdict["periodic"] = self.periodic
-        jdict["output_mesh_name"] = self.output_mesh_name
+        jdict["output_mesh_name"] = output_mesh_name
         jdict["show_viewer"] = self.show_viewer
         jdict["step_num"] = self.step_num
         jdict["show_axes"] = self.show_axes
         jdict["num_threads"] = self.num_threads
+        jdict["functional_normal_weight"] = self.fc_normal_weight
+        jdict["functional_mutate_weight"] = self.fc_mutate_weight
+        jdict["field_power"] = self.field_power
         jdict["capsules"] = []
 
         for capsule in self.capsules:
@@ -53,7 +60,13 @@ class FiberGen:
 
         json.dump(jdict, tmpfile, indent=4)
         tmpfile.close()
-        command = "{0} {1}; rm ./{1}".format(self.fibergen_path, tmpfilename)
+
+        command = None
+        if self.keepjson:
+            command = "{0} {1}; ".format(self.fibergen_path, tmpfilename)
+        else:
+            command = "{0} {1}; rm ./{1}".format(self.fibergen_path, tmpfilename)
+
         print(command)
         os.system(command)
